Delete NSString+Condensing.m
