+++ Simplenote/Simplenote-Bridging-Header.h
@@ -24,7 +24,6 @@
 #pragma mark - Extensions
 
 #import "Simperium+Simplenote.h"
-#import "NSString+Condensing.h"
 #import "NSString+Metadata.h"
 #import "NSNotification+Simplenote.h"
 #import "NSMutableAttributedString+Styling.h"
